Redone types after week2 meeting.
For example, it's not our job to detect or enforce the
hierarchy/organization of features (like the fact that you can only
specify +/-augmented if +group) -- conforming to the hierarchy is the
job of the input text.  In fact, KP's learning algorithm doesn't
currently take any advantage of the feature organization that it can
assume.  We're free to come up with a way for it to take advantage of
it.